Update the gemspec with the latest version of json-jwt
This change is to make sure xero-ruby is using the latest release of Json-jwt which will in turn allow for OpenSSL support. The current version of the library being used is over a year old and doesn't support OpenSSL 3.0 which is becoming the default on most OS distributions.

Current changes pass all tests without issue.